% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIA-helper.R
\name{combfunc}
\alias{combfunc}
\title{SPIA combfunc method.}
\usage{
combfunc(p1 = NULL, p2 = NULL, combine)
}
\arguments{
\item{p1}{See SPIA function}

\item{p2}{See SPIA function}

\item{combine}{See SPIA function}
}
\value{
See SPIA function
}
\description{
This function is combfunc from the original SPIA method.
}
