% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{combineDEAnalysisResults}
\alias{combineDEAnalysisResults}
\title{Combine DE analysis results}
\usage{
combineDEAnalysisResults(
  DEResults,
  method = c("fisher", "stouffer", "geoMean", "addCLT", "minP")
)
}
\arguments{
\item{DEResults}{A list of dataframes containing DE analysis results.
Each dataframe must have ID, p.value, logFC and logFCSE columns.}

\item{method}{The method to combine p-values. It can be one of "fisher", "stouffer", "geoMean", "addCLT", "minP".}
}
\value{
A dataframe containing combined DE analysis results.
The dataframe has ID, p.value, pDFR, logFC, and logFCSE columns.
}
\description{
This function performs mata analysis on multiple DE analysis results.
}
