% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{combinePathwayAnalysisResults}
\alias{combinePathwayAnalysisResults}
\title{Perform Meta Analysis}
\usage{
combinePathwayAnalysisResults(
  DFsList,
  method = c("fisher", "stouffer", "min", "ES", "geoMean", "addCLT", "minP")
)
}
\arguments{
\item{DFsList}{A list of dataframes obtained from runPathwayAnalysis.}
}
\value{
A dataframe of meta analysis results including combined normalized score and combined p-value for each pathway.
}
\description{
This function performs mata analysis on multiple pathway analysis results.
}
\details{
This function performs mata analysis on multiple pathway analysis results.
}
