% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{.combineEnrichmentScores}
\alias{.combineEnrichmentScores}
\title{Combine Normalized Enrichment Scores}
\usage{
.combineEnrichmentScores(inputData)
}
\arguments{
\item{inputData}{A dataframe of pathway analysis results from multiple studies.}
}
\value{
A dataframe of meta analysis results.
}
\description{
This function combines normalized enrichment scores based on REML method.
This function is used internally by combinePathwayAnalysisResults and .combinePvalues.
}
\details{
This function is used internally by combinePathwayAnalysisResults and .combinePvalues.
}
