% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{.combinePvalues}
\alias{.combinePvalues}
\title{Combine PValues}
\usage{
.combinePvalues(inputData, method = "fisher")
}
\arguments{
\item{inputData}{A dataframe of pathway analysis results from multiple studies.}

\item{method}{The selected method to combine PValues, including Fisher, Stouffer, addCLT, GeoMean, and minP.}
}
\value{
A dataframe of meta analysis results.
}
\description{
This function combines P-Values based on the selected method.
This function is used internally by combinePathwayAnalysisResults.
}
\details{
This function is used internally by combinePathwayAnalysisResults.
}
