% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO-download.R
\name{.downloadGEOObject}
\alias{.downloadGEOObject}
\title{Download GEO object}
\usage{
.downloadGEOObject(GEOID, platform, destDir)
}
\arguments{
\item{GEOID}{The GEO dataset under query, ex. GSE14762.}

\item{platform}{The platform of selected GEO dataset, ex. GPL4866.}

\item{destDir}{The user path to save the downloaded files.}
}
\value{
A GEO object with class ExpressionSet.
}
\description{
This function downloads the GEO object using getGEO function.
This function is used internally by downloadGEO.
}
\details{
This function is used internally by downloadGEO.
}
