% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO-download.R
\name{.downloadSamples}
\alias{.downloadSamples}
\title{Download samples from GEO object}
\usage{
.downloadSamples(sampleIDs, protocol, destDir)
}
\arguments{
\item{sampleIDs}{A vector of GEO accession sample IDs to be downloaded.}

\item{protocol}{The protocol of the selected GEO dataset, including affymetrix and agilent.}

\item{destDir}{The user path to save the downloaded files.}
}
\value{
TRUE
}
\description{
This function downloads the corresponding samples from queried GEO object. This function is only used if the protocol is affymetrix or agilent.
This function is used internally by downloadGEO.
}
\details{
This function is used internally by downloadGEO.
}
