% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extractPairInfo}
\alias{.extractPairInfo}
\title{extract grouping information from design matrix and contrast matrix}
\usage{
.extractPairInfo(design, contrast)
}
\arguments{
\item{design}{The design matrix.}

\item{contrast}{The contrast matrix.}
}
\value{
A list with two elements: group and pair.
group is a vector of group information.
pair is a vector of pair information.
}
\description{
This function extracts grouping information from design matrix and contrast matrix.
This function is used internally.
}
