% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{.getGOTermNames}
\alias{.getGOTermNames}
\title{Get GO terms names}
\usage{
.getGOTermNames(
  namespace = c("biological_process", "molecular_function", "cellular_component")
)
}
\arguments{
\item{namespace}{The namespace of the GO terms. E.g, biological_process, molecular_function, cellular_component.}
}
\value{
A named vector with GO terms names.
}
\description{
This function retrieves GO terms for a given organism.
}
