% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{.getGOTerms}
\alias{.getGOTerms}
\title{Get GO terms}
\usage{
.getGOTerms(
  taxid = 9606,
  namespace = c("biological_process", "molecular_function", "cellular_component")
)
}
\arguments{
\item{taxid}{The NCBI taxonomy ID of the organism.}

\item{namespace}{The namespace of the GO terms. E.g, biological_process, molecular_function, cellular_component.}
}
\value{
A named list with three elements: database, genesets and names.
}
\description{
This function retrieves GO terms for a given organism.
}
