% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID-mapping.R
\name{.getIDMappingAnnotation}
\alias{.getIDMappingAnnotation}
\title{Get ID mapping annotation from GEO platform}
\usage{
.getIDMappingAnnotation(platform)
}
\arguments{
\item{platform}{platform ID. E.g., GPL570}
}
\value{
A data frame with ID mapping annotation and two columns: FROM and TO.
The first column is the probe ID and the second column is the entrez ID.
}
\description{
This function gets ID mapping annotation from GEO platform.
This function is used internally by runDEAnalysis.
}
