% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{.getKEGGGeneSets}
\alias{.getKEGGGeneSets}
\title{Get KEGG gene sets}
\usage{
.getKEGGGeneSets(org = "hsa")
}
\arguments{
\item{org}{The organism abbreviation. E.g, hsa, mmu, dme, etc.
Visit https://www.genome.jp/kegg/catalog/org_list.html to see the full list of supported organisms.}
}
\value{
A named list with three elements: database, genesets and names.
}
\description{
This function retrieves KEGG gene sets for a given organism.
}
\examples{
\dontrun{
.getKEGGGeneSets("hsa")
}
}
