% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{.getKEGGPathwayNames}
\alias{.getKEGGPathwayNames}
\title{Get KEGG gene sets names}
\usage{
.getKEGGPathwayNames(org = "hsa")
}
\arguments{
\item{org}{The organism abbreviation. E.g, hsa, mmu, dme, etc.
Visit https://www.genome.jp/kegg/catalog/org_list.html to see the full list of supported organisms.}
}
\value{
A named vector with KEGG gene sets names.
}
\description{
This function retrieves KEGG gene sets names for a given organism.
}
\examples{
getKEGGPathwayNames("hsa")
}
