% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID-mapping.R
\name{.mapIDs}
\alias{.mapIDs}
\title{Map probe IDs to gene symbols}
\usage{
.mapIDs(exprs, annotation, DEResults)
}
\arguments{
\item{exprs}{Expression matrix. Rows are genes and columns are samples.}

\item{annotation}{Annotation data frame with mapping between probe IDs and entrez IDs.
The data frame must have two columns: FROM and TO. The first column is the probe ID and the second column is the entrez ID.}

\item{DEResults}{DE analysis results data frame. Must have a column named ID and p.value.}
}
\value{
A list with two elements:
\itemize{
 \item exprs: Expression matrix with probe IDs mapped to gene symbols. Rows are genes and columns are samples.
 \item DEResults: DE analysis results data frame with probe IDs mapped to gene symbols.
}
}
\description{
This function maps probe IDs to gene symbols.
This function is used internally by runDEAnalysis.
}
