% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID-mapping.R
\name{.mapProbeIDsToGeneSymbols}
\alias{.mapProbeIDsToGeneSymbols}
\title{Map probe IDs to gene symbols}
\usage{
.mapProbeIDsToGeneSymbols(exprs, annotation, DEResults)
}
\arguments{
\item{exprs}{Expression matrix. Rows are genes and columns are samples.}

\item{annotation}{Annotation data frame with mapping between probe IDs and gene symbols. The first column is the probe ID and the second column is the gene symbol.}

\item{DEResults}{DE analysis results data frame. Must have a column named ID and p.value.}

\item{mappingMethod}{Method to use for mapping probe IDs to gene symbols. Can be "mean", "max", "median", "mostSignificant".}
}
\value{
A list with two elements:
\itemize{
 \item exprs: Expression matrix with probe IDs mapped to gene symbols. Rows are genes and columns are samples.
 \item DEResults: DE analysis results data frame with probe IDs mapped to gene symbols.
}
}
\description{
This function maps probe IDs to gene symbols.
This function is used internally by runDEAnalysis.
}
