% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO-download.R
\name{.processAffymetrix}
\alias{.processAffymetrix}
\title{Process and normalize affymetrix-based dataset}
\usage{
.processAffymetrix(metadata, sampleIDs, destDir)
}
\arguments{
\item{metadata}{The metadat of downloaded GEO object.}

\item{sampleIDs}{A vector of downloaded samples IDs from the queried GEO object.}

\item{destDir}{The user path to save the downloaded files.}
}
\value{
A SummarizedExperiment object with appended expression values as assay and metadata as colData.
}
\description{
This function normalize expression data of the downloaded GEO object.
The normalization process includes background correction using RMA and data normalization using quantile method.
This function is only used if the protocol is affymetrix or agilent.
This function is used internally by downloadGEO.
}
\details{
This function is used internally by downloadGEO.
}
