% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO-download.R
\name{.processAgilent}
\alias{.processAgilent}
\title{Process and normalize agilent-based dataset}
\usage{
.processAgilent(metadata, sampleIDs, destDir)
}
\arguments{
\item{metadata}{The metadat of downloaded GEO object.}

\item{sampleIDs}{A vector of downloaded samples IDs from the queried GEO object.}

\item{destDir}{The user path to save the downloaded files.}
}
\value{
A SummarizedExperiment object with appended expression values as assay and metadata as colData.
}
\description{
This function normalize expression data of the downloaded GEO object.
In the normalization process the limma package functions are utilized, including background correction using normexp and data normalization using loess and quantile methods.
This function is only used if the protocol is agilent.
This function is used internally by downloadGEO.
}
\details{
This function is used internally by downloadGEO.
}
