% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.requirePackage}
\alias{.requirePackage}
\title{Check if a package is installed and install it if not}
\usage{
.requirePackage(pkg)
}
\arguments{
\item{pkg}{The name of the package to be loaded.}
}
\value{
TRUE if the package is installed, FALSE otherwise.
}
\description{
This function checks if a package is installed and install it if not.
This function is internally used by the package.
}
