% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-analysis.R
\name{.runCePaGSA}
\alias{.runCePaGSA}
\title{Pathway Enrichment Analysis using CePaGSA}
\usage{
.runCePaGSA(summarizedExperiment, network, ...)
}
\arguments{
\item{summarizedExperiment}{The generated SummarizedExpriment object from DE analysis result.}

\item{network}{The pathways network catalouge generated by getCePaPathwayCatalogue.}

\item{...}{A list of other passed arguments to GSA. See GSA function.}
}
\value{
A dataframe of pathway analysis results
}
\description{
This function performs patwhay analysis using CePaGSA.
This function is used internally by runPathwayAnalysis.
}
\details{
This function is used internally by runPathwayAnalysis.
}
