% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-analysis.R
\name{.runCePaORA}
\alias{.runCePaORA}
\title{Pathway Enrichment Analysis using CePaORA}
\usage{
.runCePaORA(summarizedExperiment, network, pThreshold, bk, ...)
}
\arguments{
\item{summarizedExperiment}{The generated SummarizedExpriment object from DE analysis result.}

\item{network}{The pathways network catalouge generated by getCePaPathwayCatalogue.}

\item{pThreshold}{The p.value cutoff to select differentially expressed genes.}

\item{bk}{A vector contain of DE genes.}

\item{...}{A list of other passed arguments to ORA. See ORA function.}
}
\value{
A dataframe of pathway analysis results
}
\description{
This function performs patwhay analysis using CePaORA method.
This function is used internally by runPathwayAnalysis.
}
\details{
This function is used internally by runPathwayAnalysis.
}
