% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE-analysis.R
\name{.runDESeq2}
\alias{.runDESeq2}
\title{Differential expression analysis using DESeq2}
\usage{
.runDESeq2(exprs, controlSamples, conditionSamples)
}
\arguments{
\item{exprs}{Raw counts matrix. Rows are genes and columns are samples.}

\item{controlSamples}{Vector of control samples.}

\item{conditionSamples}{Vector of condition samples.}
}
\value{
A data frame with DE analysis results
}
\description{
This function performs differential expression analysis using limma.
This function is used internally by runDEAnalysis.
}
