% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-analysis.R
\name{.runFgsea}
\alias{.runFgsea}
\title{Pathway Enrichment Analysis using fgsea}
\usage{
.runFgsea(summarizedExprimentObj, gs, nperm)
}
\arguments{
\item{summarizedExprimentObj}{The generated SummarizedExpriment object from DE analysis result.}

\item{nperm}{The number of permutations to run pathway analysis.}

\item{genesets}{The genesets definition, ex. KEGG genesets.}
}
\value{
A dataframe of pathway analysis results
}
\description{
This function performs pathway analysis using fgsea (fast geneset enrichment analysis).
This function is used internally by runPathwayAnalysis.
}
\details{
This function is used internally by runPathwayAnalysis.
}
