% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{.runFisher}
\alias{.runFisher}
\title{Combine P-Valuse using Fisher}
\usage{
.runFisher(pvals)
}
\arguments{
\item{pvals}{The vector of P-Values to be combined.}
}
\value{
A combined P-Value
}
\description{
This function combines P-Values based on Fisher method.
This function is used internally by .combinePvalues.
}
\details{
This function is used internally by .combinePvalues.
}
