% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{.runGeoMean}
\alias{.runGeoMean}
\title{Combine P-Valuse using Geometric Mean}
\usage{
.runGeoMean(pvals)
}
\arguments{
\item{pvals}{The vector of P-Values to be combined.}
}
\value{
A combined P-Value
}
\description{
This function combines P-Values by computing their geometric mean.
This function is used internally by .combinePvalues.
}
\details{
This function is used internally by .combinePvalues.
}
