% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset-analysis.R
\name{.runKsWilcox}
\alias{.runKsWilcox}
\title{Geneset Enrichment Analysis using KS or Wilcox}
\usage{
.runKsWilcox(summarizedExperiment, genesets, sTest)
}
\arguments{
\item{summarizedExperiment}{The generated SummarizedExpriment object from DE analysis result.}

\item{genesets}{The genesets definition, ex. KEGG genesets from getGeneSets function.}

\item{sTest}{The selected test to perform geneset analysis, either 'ks' or 'wilcox'.}
}
\value{
A dataframe of geneset analysis results
}
\description{
This function performs geneset analysis using KS or Wilcox test.
This function is used internally by runGeneSetEnrichmentAnalysis.
}
\details{
This function is used internally by runGeneSetEnrichmentAnalysis.
}
