% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-analysis.R
\name{.runORA}
\alias{.runORA}
\title{Pathway Enrichment Analysis using ORA}
\usage{
.runORA(summarizedExperiment, genesets, pThreshold = 0.05)
}
\arguments{
\item{summarizedExperiment}{The generated SummarizedExpriment object from runDEAnalysis.}

\item{genesets}{The genesets definition, ex. KEGG genesets.}
}
\value{
A dataframe of pathway analysis results
}
\description{
This function performs pathway analysis based on ORA (Over Representation Analysis).
This function is used internally by runPathwayAnalysis.
}
\details{
This function is used internally by runPathwayAnalysis.
}
