% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-analysis.R
\name{.runSPIA}
\alias{.runSPIA}
\title{Pathway Enrichment Analysis using SPIA}
\usage{
.runSPIA(summarizedExperiment, network, pThreshold, all, ...)
}
\arguments{
\item{summarizedExperiment}{The generated SummarizedExpriment object from DE analysis result.}

\item{network}{The network definition generated by getSPIAKEGGNetwork method.}

\item{pThreshold}{The p.value cutoff to select differentially expressed genes.}

\item{all}{The spia argument. See spia function.}

\item{...}{A list of other passed arguments to spia. See spia function.}
}
\value{
A dataframe of pathway analysis results
}
\description{
This function performs patwhay analysis using SPIA method.
This function is used internally by runPathwayAnalysis.
}
\details{
This function is used internally by runPathwayAnalysis.
}
