% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-analysis.R
\name{.runStouffer}
\alias{.runStouffer}
\title{Combine P-Valuse using Stouffer}
\usage{
.runStouffer(pvals)
}
\arguments{
\item{pvals}{The vector of P-Values to be combined.}
}
\value{
A combined P-Value
}
\description{
This function combines P-Values based on Stouffer method.
This function is used internally by .combinePvalues.
}
\details{
This function is used internally by .combinePvalues.
}
