% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE-analysis.R
\name{.runTTest}
\alias{.runTTest}
\title{Differential expression analysis using t-test}
\usage{
.runTTest(exprs, controlSamples, conditionSamples, isPaired)
}
\arguments{
\item{exprs}{Normalized expression matrix. Rows are genes and columns are samples.}

\item{controlSamples}{Vector of control samples.}

\item{conditionSamples}{Vector of condition samples.}
}
\value{
A data frame with DE analysis results
}
\description{
This function performs differential expression analysis using t-test.
This function is used internally by runDEAnalysis.
}
