% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIA-helper.R
\name{.SPIAMod}
\alias{.SPIAMod}
\title{SPIA method modified with inputs are KEGG pathway networks instead of a folder.}
\usage{
.SPIAMod(
  de = NULL,
  all = NULL,
  pathInfo,
  nB = 2000,
  verbose = TRUE,
  beta = NULL,
  combine = "fisher"
)
}
\arguments{
\item{de}{See SPIA function}

\item{all}{See SPIA function}

\item{pathInfo}{pathway information generated by getSPIAKEGGNetwork function}

\item{nB}{See SPIA function}

\item{verbose}{See SPIA function}

\item{beta}{See SPIA function}

\item{combine}{See SPIA function}
}
\value{
See SPIA function
}
\description{
This function is modified from the original SPIA method to accept KEGG pathway networks as inputs.
}
