% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CePa-helper.R
\name{getCePaPathwayCatalogue}
\alias{getCePaPathwayCatalogue}
\title{Get KEGG pathway catalouge (network) for CePa.ORA and CePa.GSA methods}
\usage{
getCePaPathwayCatalogue(org = "hsa", updateCache = F)
}
\arguments{
\item{org}{The organism abbreviation. E.g, hsa, mmu, dme, etc.
Visit https://www.genome.jp/kegg/catalog/org_list.html to see the full list of supported organisms.}
}
\value{
A named list of KEGG pathways catalouge for CePa.ORA and CePa.GSA methods.
}
\description{
Get KEGG pathway catalouge for CePa.ORA and CePa.GSA methods
}
\examples{
getCePaPathwayCatalogue("hsa")
}
