% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID-mapping.R
\name{getSupportedPlatforms}
\alias{getSupportedPlatforms}
\title{Get supported platforms}
\usage{
getSupportedPlatforms()
}
\value{
A named list of supported platforms, where the names are the platform IDs and the values are the corresponding annotation packages from Bioconductor.
}
\description{
Get supported platforms
}
\examples{

library(RCPA)
supportedPlatforms <- getSupportedPlatforms()

}
