% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadData}
\alias{loadData}
\title{Load data from GitHub}
\usage{
loadData(name)
}
\arguments{
\item{name}{The name of the data.}
}
\value{
Load the data with the specified name.
}
\description{
This function loads data from GitHub.
}
\examples{

library(RCPA)
RNASeqDataset <- loadData("RNASeqDataset")


}
