% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gene-heatmap.R
\name{plotDEGeneHeatmap}
\alias{plotDEGeneHeatmap}
\title{Plot gene heatmap from a SummarizedExperiment object with DE analysis results}
\usage{
plotDEGeneHeatmap(
  DEResults,
  genes,
  useFDR = TRUE,
  labels = NULL,
  logFCLims = c(-5, 5),
  negLog10pValueLims = c(0, 5)
)
}
\arguments{
\item{DEResults}{A named list of data frame of DE analysis results.}

\item{genes}{A vector of gene id (e.g. Entrez IDs) to plot.
The genes must be in the ID column of the data frame in DEResults.}

\item{useFDR}{If TRUE, use FDR adjusted p-values. Otherwise, use raw p-values.}

\item{labels}{A vector of labels for the genes. If not provided, the gene IDs will be used as labels.}

\item{logFCLims}{A vector of length 2 specifying the minimum and maximum log fold change to plot.}

\item{negLog10pValueLims}{A vector of length 2 specifying the minimum and maximum -log10(p-value) to plot.}
}
\value{
A heatmap of the genes from ggplot2.
}
\description{
Plot gene heatmap from a SummarizedExperiment object with DE analysis results.
The heatmap contains p-values and log fold changes from the DE analysis.
}
\examples{
\dontrun{
#Load necessary libraries
library(SummarizedExperiment)
library(limma)
library(RCPA)
library(ggplot2)
# Load dataset
data("data")
# Get affymetrix dataset
affyDataset <- data$affyDataset
# Create the analysis design
affyDesign <- model.matrix(~0 + condition + region, data = colData(affyDataset))
affyContrast <- limma::makeContrasts("conditionalzheimer-conditionnormal", levels=affyDesign)
# Perform DE analysis affymetrix dataset
affyDEExperiment <- RCPA::runDEAnalysis(affyDataset,
                                        method = "limma",
                                        design = affyDesign,
                                        contrast = affyContrast,
                                        annotation = "GPL570")

# Get Agilent dataset
agilDataset <- data$agilDataset
# Create the analysis design
agilDesign <- model.matrix(~0 + condition, data = colData(agilDataset))
agilContrast <- limma::makeContrasts(conditionalzheimer-conditionnormal, levels=agilDesign)
# Perform genID mapping
GPL4133Anno <- GEOquery::dataTable(GEOquery::getGEO("GPL4133"))@table
GPL4133GeneMapping <- data.frame(FROM = GPL4133Anno$SPOT_ID, 
                                 TO = as.character(GPL4133Anno$GENE), 
                                 stringsAsFactors = F)
GPL4133GeneMapping <- GPL4133GeneMapping[!is.na(GPL4133GeneMapping$TO), ]
# Perform DE analysis for agilent dataset
agilDEExperiment <- RCPA::runDEAnalysis(agilDataset, 
                                        method = "limma", 
                                        design = agilDesign, 
                                        contrast = agilContrast, 
                                        annotation = GPL4133GeneMapping)


rest <- list(affyDE = rowData(affyDEExperiment), agilDE = rowData(agilDEExperiment))

# Get common genes from different studies
genes <- sample(Reduce(intersect, lapply(rest, function(res) res$ID)), 30)
# Get gene annotation
annotation <- getEntrezAnnotation(genes)
# Get gene description
labels <- annotation[genes, "Description"]
# Plot gene heatmap
plotDEGeneHeatmap(rest, genes, labels = labels)
}
}
