% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-forest.R
\name{plotForest}
\alias{plotForest}
\title{Plot forest plot from pathway/geneset/meta analysis results}
\usage{
plotForest(
  resultsList,
  yAxis = c("ID", "name"),
  axisFontSize = 10,
  titleFontSize = 12,
  pointSize = 7
)
}
\arguments{
\item{resultsList}{A named list of dataframes from pathway analysis, geneset analysis, and/or meta analysis results.
The columns are ID, name, description, p.value, pFDR, size, nDE, score and normalizedScore.}

\item{yAxis}{The column to use for the y-axis.}

\item{axisFontSize}{The font size of axis labels.}

\item{titleFontSize}{The font size for plot title.}

\item{pointSize}{The red point size in the plot.}
}
\value{
A grid of one or more forest plots from grid.arrange.
}
\description{
Plot forest plot from pathway/geneset/meta analysis results.
}
\examples{
#TODO add example
}
