% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-KEGG-map.R
\name{plotKEGGMap}
\alias{plotKEGGMap}
\title{Plot KEGG map with DE genes}
\usage{
plotKEGGMap(
  DEResults,
  KEGGPathwayID,
  statistic = "logFC",
  useFDR = TRUE,
  pThreshold = 0.05,
  statLimit = 3
)
}
\arguments{
\item{DEResults}{A named list of data frame of DE analysis results.
The columns of each data frame should be at least ID, logFC, p.value and pFDR.}

\item{KEGGPathwayID}{The KEGG pathway ID.}

\item{statistic}{The column name of the statistic used to plot the DE genes.
If statistic is p.value or pFDR, all genes are colored.
Otherwise, only DE genes are colored.}

\item{useFDR}{If TRUE, DE genes are selected based on pFDR, otherwise p.value.}

\item{pThreshold}{The p-value threshold to select DE genes.
Only used when statistic is not p.value or pFDR.}

\item{statLimit}{The absolute value of the statistic to color the DE genes.
If statistic is p.value or pFDR, this parameter is the limit of -log10(p-value).
Otherwise, this parameter is the limit of the absolute value of the statistic.}
}
\value{
A list with the following elements:
\itemize{
\item plot: A ggplot object of the KEGG map.
\item width: The width of the KEGG map.
\item height: The height of the KEGG map.
}
}
\description{
This function plots KEGG map with DE genes.
}
\examples{
\dontrun{
#Load necessary libraries
library(SummarizedExperiment)
library(limma)
library(RCPA)
library(ggplot2)
# Load dataset
data("data")
# Get affymetrix dataset
affyDataset <- data$affyDataset
# Create the analysis design
affyDesign <- model.matrix(~0 + condition + region, data = colData(affyDataset))
affyContrast <- limma::makeContrasts("conditionalzheimer-conditionnormal", levels=affyDesign)
# Perform DE analysis affymetrix dataset
affyDEExperiment <- RCPA::runDEAnalysis(affyDataset,
                                        method = "limma",
                                        design = affyDesign,
                                        contrast = affyContrast,
                                        annotation = "GPL570")
# Get DE analysis result
rest <- list(rowData(affyDEExperiment))
# Plot KEGG pathways heatmap using log fold change statistics
plotKEGGMap(rest, "path:hsa05200", statistic = "logFC", useFDR = TRUE, pThreshold = 0.05)
# Plot KEGG pathways heatmap using p-value statistics
plotKEGGMap(rest, "path:hsa05200", statistic = "p.value", useFDR = FALSE, pThreshold = 0.05)
}
}
