% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pathway-heatmap.R
\name{plotPathwayHeatmap}
\alias{plotPathwayHeatmap}
\title{Plot pathways heatmap plot from pathway/geneset/meta analysis results}
\usage{
plotPathwayHeatmap(
  resultsList,
  yAxis = c("ID", "name"),
  negLog10pValueLims = c(0, 5)
)
}
\arguments{
\item{resultsList}{A named list of dataframes from pathway analysis, geneset analysis, and/or meta analysis results.
The columns are ID, name, description, p.value, pFDR, size, nDE, score and normalizedScore.}

\item{yAxis}{The column to use for the y-axis.}

\item{negLog10pValueLims}{A vector of length 2 specifying the minimum and maximum -log10(p-value) to plot.}
}
\value{
A ggplot2 object for presenting the heatmap of the pathways.
}
\description{
pathways heatmap forest plot from pathway/geneset/meta analysis results.
}
\examples{
#TODO add example
}
