% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-venn.R
\name{plotVennPathway}
\alias{plotVennPathway}
\title{Plot venndiagram from multiple pathway analysis results}
\usage{
plotVennPathway(PAResults, pThreshold = 0.05, useFDR = TRUE)
}
\arguments{
\item{PAResults}{A list of data frames with the results of pathway analysis.}

\item{pThreshold}{The p-value threshold to determine if a pathway is enriched.}

\item{useFDR}{Use the FDR adjusted p-value instead of the raw p-value.}
}
\description{
Plot a venndiagram from multiple pathway analysis results.
}
\examples{
\dontrun{
library(RCPA)

loadData("affyDEExperiment")
loadData("agilDEExperiment")
loadData("RNASeqDEExperiment")

DEResults <- list(
    "Affymetrix - GSE5281" = rowData(affyDEExperiment),
    "Agilent - GSE61196"   = rowData(agilDEExperiment),
    "RNASeq - GSE153873"   = rowData(RNASeqDEExperiment)
)
DEResultUps <- lapply(DEResults, function(df) df[!is.na(df$logFC) & df$logFC > 0, ])
DEResultDowns <- lapply(DEResults, function(df) df[!is.na(df$logFC) & df$logFC < 0, ])

RCPA::plotVennDE(DEResults) + ggplot2::ggtitle("Differential Expression Genes")
RCPA::plotVennDE(DEResultUps) + ggplot2::ggtitle("Up-regulated Genes")
RCPA::plotVennDE(DEResultDowns) + ggplot2::ggtitle("Down-regulated Genes")

}
}
