% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{plotVolcanoPathway}
\alias{plotVolcanoPathway}
\title{Plot volcano plot from Pathway analysis results}
\usage{
plotVolcanoPathway(
  PAResult,
  xAxis = c("normalizedScore", "score"),
  yAxis = c("-log10(pFDR)", "-log10(p.value)"),
  pThreshold = 0.05,
  label = "name",
  IDsToLabel = NULL,
  topToLabel = 10,
  sideToLabel = c("both", "left", "right")
)
}
\arguments{
\item{PAResult}{A data frame with Pathway analysis results.
The columns are ID, name, description, p.value, pFDR, pathwaySize, nDE, score and normalizedScore.}

\item{xAxis}{The column to use for the x-axis.}

\item{yAxis}{The column to use for the y-axis.}

\item{pThreshold}{The p-value threshold to use for the horizontal line.}

\item{label}{The column to use for the labels. Default is "name".}

\item{IDsToLabel}{A vector of IDs to label.
When NULL, the top pathways are labeled. Default is NULL.}

\item{topToLabel}{The number of top pathways to label when IDsToLabels is NULL.}
}
\description{
Plot volcano plot from Pathway analysis results
}
\examples{
\dontrun{
library(AnnotationDbi)
library(SummarizedExperiment)
library(limma)
library(RCPA)
data("data")
# Get affymetrix dataset
affyDataset <- data$affyDataset
# Create the analysis design
affyDesign <- model.matrix(~0 + condition + region, data = colData(affyDataset))
affyContrast <- limma::makeContrasts("conditionalzheimer-conditionnormal", levels=affyDesign)
# Perform DE analysis affymetrix dataset
affyDEExperiment <- RCPA::runDEAnalysis(affyDataset, 
                                        method = "limma", 
                                        design = affyDesign, 
                                        contrast = affyContrast, 
                                        annotation = "GPL570")
# Get DE data frame
DEResult <- rowData(affyDEExperiment)
# Plot volcano
plotVolcanoDE(DEResult)
}

}
