% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus-analysis.R
\name{runConsensusAnalysis}
\alias{runConsensusAnalysis}
\title{Consensus Analysis}
\usage{
runConsensusAnalysis(
  PAResults,
  method = c("weighted.mean", "RRA", "min", "geom.mean", "mean", "median", "stuart"),
  weightsList = NULL,
  rank.by = c("normalizedScore", "pFDR", "both"),
  backgroundSpace = NULL
)
}
\arguments{
\item{PAResults}{A list of at least length two from enrichment analysis results.}

\item{method}{The consensus analsyis method.}

\item{weightsList}{A vector of integer values.
Each element shows the corresponding input result weight.
When selected method is weightedAvg this parameter needs to be specified.
If it is null all the weights are considered as equal.}

\item{rank.by}{An string parameter which specifies how the input results should be ranked.}

\item{backgroundSpace}{A list of lists with the same length as PAResults.
Each list contains underlying space (set of pathways) for each input dataframe in PAResults.
This parameter is optional.
NULL means all input dataframes share a common space.
So the union pathways of all input dataframes is taken into account.}
}
\value{
A dataframe of consensus analysis result
}
\description{
This function performs consensus analysis using several methods.
These methods are: weightedAvg, RRA, min, geom.mean, mean, median, and stuart.
}
\examples{
\donttest{

library(RCPA)

}
}
