% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE-analysis.R
\name{runDEAnalysis}
\alias{runDEAnalysis}
\title{Differential expression analysis}
\usage{
runDEAnalysis(
  summarizedExperiment,
  method = c("limma", "DESeq2", "edgeR"),
  design,
  contrast,
  annotation = NULL
)
}
\arguments{
\item{summarizedExperiment}{SummarizedExperiment object}

\item{method}{Method to use for differential expression analysis. Can be "limma", "DESeq2" or "edgeR".}

\item{annotation}{A data frame, ChipDb, or OrgDb object with mapping between probe IDs and gene symbols.
If not provided, the function will try to get the mapping from the platform annotation in the SummarizedExperiment object.
If the annotation is not available, the function will return the probe IDs.
Regardless of the type of annotation, it must contains two columns: PROBEID and SYMBOL.}

\item{controlSamples}{Vector of control samples. Each sample is a column in the SummarizedExperiment object.}

\item{conditionSamples}{Vector of condition samples. Each sample is a column in the SummarizedExperiment object.}
}
\value{
A SummarizedExperiment object with DE analysis results appended to the rowData slot with the following columns:
\itemize{
\item{logFC}{log2 fold change}
\item{p.value}{p-value}
\item{pFDR}{p-value adjusted for multiple testing using Benjamini-Hochberg method}
\item{control.mean}{mean expression in control samples}
\item{condition.mean}{mean expression in condition samples}
}
The assay slot will contain only samples from the control and condition vectors,
and the rownames will be the gene symbols.
Other slots will be the same as in the input SummarizedExperiment object.
}
\description{
This function performs differential expression analysis using either limma, DESeq2 or edgeR.
}
