% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset-analysis.R
\name{runGeneSetEnrichmentAnalysis}
\alias{runGeneSetEnrichmentAnalysis}
\title{Geneset Enrichment Analysis}
\usage{
runGeneSetEnrichmentAnalysis(
  summarizedExperiment,
  genesets,
  method = c("ora", "fgsea", "gsa", "ks", "wilcox"),
  ORAArgs = list(pThreshold = 0.05),
  FgseaArgs = list(sampleSize = 101, minSize = 1, maxSize = Inf, eps = 1e-50, scoreType =
    "std", nproc = 0, gseaParam = 1, BPPARAM = NULL, nPermSimple = 1000, absEps = NULL),
  GSAArgs = list(method = "maxmean", random.seed = NULL, knn.neighbors = 10, s0 = NULL,
    s0.perc = NULL, minsize = 15, maxsize = 500, restand = TRUE, restand.basis =
    "catalog", nperms = 200, xl.mode = "regular", xl.time = NULL, xl.prevfit = NULL)
)
}
\arguments{
\item{summarizedExperiment}{The generated SummarizedExpriment object from DE analysis result.}

\item{genesets}{The genesets definition, ex. KEGG genesets from getGeneSets function.}

\item{method}{The geneset analsyis method, including ORA, fgsea, GSA, ks, and wilcox.}

\item{ORAArgs}{A list of other passed arguments to ORA. pThreshold is used as p.value cutoff to pick DE genes.}

\item{FgseaArgs}{A list of other passed arguments to fgsea. See fgsea function.}

\item{GSAArgs}{A list of other passed arguments to GSA. See GSA function.}
}
\value{
A dataframe of geneset analysis result
}
\description{
This function performs geneset analysis using either ORA, fgsea, GSA, ks, or wilcox approaches.
}
\examples{
\dontrun{
# Loading libraries
library(SummarizedExperiment)
library(limma)
library(RCPA)
data("data")
# Get affymetrix dataset
affyDataset <- data$affyDataset
# Create the analysis design
affyDesign <- model.matrix(~0 + condition + region, data = colData(affyDataset))
affyContrast <- limma::makeContrasts("conditionalzheimer-conditionnormal", levels=affyDesign)
# Perform DE analysis affymetrix dataset
affyDEExperiment <- RCPA::runDEAnalysis(affyDataset, 
                                        method = "limma", 
                                        design = affyDesign, 
                                        contrast = affyContrast, 
                                        annotation = "GPL570")

genesets <- getGeneSets("KEGG", "hsa", minSize = 10, maxSize = 1000)

# Perform geneset enrichment analysis using wilcox test
result <- runGeneSetEnrichmentAnalysis(affyDEExperiment, genesets, method = "wilcox")
# Perform geneset enrichment analysis using over representation analysis
result <- runGeneSetEnrichmentAnalysis(affyDEExperiment, genesets, method = "ora")
# Perform geneset enrichment analysis using fgsea
result <- runGeneSetEnrichmentAnalysis(affyDEExperiment, genesets, method = "fgsea")
# Perform geneset enrichment analysis using ks test
result <- runGeneSetEnrichmentAnalysis(affyDEExperiment, genesets, method = "ks")
# Perform geneset enrichment analysis using gsa
result <- runGeneSetEnrichmentAnalysis(affyDEExperiment, genesets, method = "gsa")
}
}
