% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID-mapping.R
\name{.getIDMappingAnnotation}
\alias{.getIDMappingAnnotation}
\title{Get ID mapping annotation from GEO platform}
\usage{
.getIDMappingAnnotation(platform)
}
\arguments{
\item{GEO}{platform ID. E.g., GPL570}

\item{dbObj}{A database object from AnnotationDbi package.}
}
\value{
A data frame with ID mapping annotation. The first column is the probe ID and the second column is the gene symbol.
}
\description{
This function gets ID mapping annotation from GEO platform.
This function is used internally by runDEAnalysis.
}
