% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset-analysis.R
\name{.runFgsea}
\alias{.runFgsea}
\title{Geneset Enrichment Analysis using fgsea}
\usage{
.runFgsea(summarizedExperiment, genesets, ...)
}
\arguments{
\item{summarizedExperiment}{The generated SummarizedExpriment object from DE analysis result.}

\item{genesets}{The genesets definition, ex. KEGG genesets from getGeneSets function.}

\item{...}{A list of other passed arguments to fgsea. See fgsea function.}
}
\value{
A dataframe of geneset analysis results
}
\description{
This function performs geneset analysis using fgsea (fast geneset enrichment analysis).
This function is used internally by runGeneSetEnrichmentAnalysis.
}
\details{
This function is used internally by runGeneSetEnrichmentAnalysis.
}
