% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset-analysis.R
\name{.runORA}
\alias{.runORA}
\title{Geneset Enrichment Analysis using ORA}
\usage{
.runORA(summarizedExperiment, genesets, pThreshold)
}
\arguments{
\item{summarizedExperiment}{The generated SummarizedExpriment object from DE analysis result.}

\item{genesets}{The genesets definition, ex. KEGG genesets from getGeneSets function.}

\item{pThreshold}{The p.value cutoff threashold.}
}
\value{
A dataframe of geneset analysis results
}
\description{
This function performs geneset analysis based on ORA (Over Representation Analysis).
This function is used internally by runGeneSetEnrichmentAnalysis.
}
\details{
This function is used internally by runGeneSetEnrichmentAnalysis.
}
