% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO-download.R
\name{downloadGEO}
\alias{downloadGEO}
\title{Download GEO data}
\usage{
downloadGEO(GEOID, platform, protocol = c("affymetrix", "agilent"), destDir)
}
\arguments{
\item{GEOID}{The ID of the GEO dataset}

\item{platform}{The platform of selected GEO dataset}

\item{protocol}{The protocol of the selected GEO dataset, including affymetrix and agilent.}
}
\value{
A SummarizedExperiment object including the processed data
}
\description{
This function download and process data from GEO for microarray and RNASeq data.
}
\examples{
downloadGEO("GSE20153", "GPL570", "affymetrix", getwd())
}
