% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIA-helper.R
\name{getSPIAKEGGNetwork}
\alias{getSPIAKEGGNetwork}
\title{Get KEGG pathway network for SPIA method}
\usage{
getSPIAKEGGNetwork(org = "hsa", updateCache = F)
}
\arguments{
\item{org}{The organism abbreviation. E.g, hsa, mmu, dme, etc.}

\item{updateCache}{A parameter to disable/enable cache update.
Visit https://www.genome.jp/kegg/catalog/org_list.html to see the full list of supported organisms.}
}
\value{
A named list of KEGG pathway networks for SPIA method.
}
\description{
Get KEGG pathway network for SPIA method
}
\examples{
\dontrun{
getSPIAKEGGNetwork("hsa")
}
}
