% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pathway-bar-chart.R
\name{plotBarChart}
\alias{plotBarChart}
\title{Plot a bar chart of the pathway analysis results}
\usage{
plotBarChart(
  results,
  limit = 50,
  label = "name",
  by = c("pFDR", "p.value", "score", "normalizedScore"),
  maxNegLog10PValue = 5,
  pThreshold = 0.05,
  useFDR = TRUE
)
}
\arguments{
\item{results}{A data frame with pathway analysis results.
The columns are ID, name, description, p.value, pFDR, size, nDE, score and normalizedScore.}

\item{limit}{The maximum number of pathways to plot.}

\item{label}{The column to use for the labels.}

\item{by}{The column to use for the bar heights.}

\item{maxNegLog10PValue}{The maximum -log10(p-value) to plot.}

\item{pThreshold}{The p-value threshold to use for significance.}

\item{useFDR}{If TRUE, use FDR adjusted p-values. Otherwise, use raw p-values.}
}
\value{
A ggplot2 object.
}
\description{
This function plots a bar chart of the pathway analysis results.
}
\examples{
#TODO add example
}
