% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-gene-heatmap.R
\name{plotDEGeneHeatmap}
\alias{plotDEGeneHeatmap}
\title{Plot gene heatmap from a SummarizedExperiment object with DE analysis results}
\usage{
plotDEGeneHeatmap(
  DEResults,
  genes,
  useFDR = TRUE,
  labels = NULL,
  logFCLims = c(-5, 5),
  negLog10pValueLims = c(0, 5)
)
}
\arguments{
\item{DEResults}{A named list of data frame of DE analysis results.}

\item{genes}{A vector of gene id (e.g. Entrez IDs) to plot.
The genes must be in the ID column of the data frame in DEResults.}

\item{useFDR}{If TRUE, use FDR adjusted p-values. Otherwise, use raw p-values.}

\item{labels}{A vector of labels for the genes. If not provided, the gene IDs will be used as labels.}

\item{logFCLims}{A vector of length 2 specifying the minimum and maximum log fold change to plot.}

\item{negLog10pValueLims}{A vector of length 2 specifying the minimum and maximum -log10(p-value) to plot.}
}
\value{
A heatmap of the genes from ggplot2.
}
\description{
Plot gene heatmap from a SummarizedExperiment object with DE analysis results.
The heatmap contains p-values and log fold changes from the DE analysis.
}
\examples{
\dontrun{
library(hgu133plus2.db)
library(AnnotationDbi)
library(SummarizedExperiment)
library(limma)
library(ggplot2)
library(RCPA)

# generate diferential expression analysis using a random gene expression matrix
DEResults <- lapply(1:3, function(seed) {
  set.seed(seed)
  exprs <- round(matrix(abs(rnorm(20000 * 10, sd = 4)), nrow = 20000, ncol = 10))
  rownames(exprs) <- sample(keys(hgu133plus2.db, keytype = "PROBEID"), nrow(exprs), replace = FALSE)
  colnames(exprs) <- paste0("sample", 1:10)
  
  controlSamples <- paste0("sample", 1:5)
  conditionSamples <- paste0("sample", 6:10)
  
  exprs[, conditionSamples] <- exprs[, conditionSamples] + 
  2*sample(c(1,-1), nrow(exprs), replace = TRUE)
  
  colData <- data.frame(
    row.names = colnames(exprs),
    group = factor(c(rep("control", length(controlSamples)), rep("condition", 
    length(conditionSamples)))),
    pair = factor(c(seq_along(controlSamples), seq_along(conditionSamples)))
  )
  
  summarizedExperiment <- SummarizedExperiment(
    assays = list(counts = exprs),
   colData = colData
)
 
 # control vs condition
 design <- model.matrix(~0 + group, data = colData)
 contrast <- makeContrasts("groupcondition-groupcontrol", levels = design)
 runDEAnalysis(summarizedExperiment,
  method = "limma", 
  design, 
  contrast, 
  annotation = "GPL570") \%>\% rowData()
})

# Get common genes from different studies
genes <- sample(Reduce(intersect, lapply(DEResults, function(res) res$ID)), 30)
# Get gene annotation
annotation <- getEntrezAnnotation(genes)
# Get gene description
labels <- annotation[genes, "Description"]
# Plot gene heatmap
plotDEGeneHeatmap(DEResults, genes, labels = labels)
}
}
