% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-forest.R
\name{plotForest}
\alias{plotForest}
\title{Plot forest plot from pathway/geneset/meta analysis results}
\usage{
plotForest(
  resultsList,
  yAxis = c("ID", "name"),
  statLims = c(-2.5, 2.5),
  useFDR = TRUE
)
}
\arguments{
\item{resultsList}{A named list of dataframes from pathway analysis, geneset analysis, and/or meta analysis results.
The columns are ID, name, description, p.value, pFDR, size, nDE, score and normalizedScore.}

\item{yAxis}{The column to use for the y-axis.}

\item{statLims}{A vector of length 2 specifying the limits for score to use in the x axis.}

\item{useFDR}{A boolean parameter to specify if adjusted p-value should be considered.}
}
\value{
A list of ggplot objects.
}
\description{
Plot forest plot from pathway/geneset/meta analysis results.
}
\examples{
\dontrun{
# Loading the library
library(RCPA)
library(ggplot2)
# Set seed for repro
set.seed(123)
# Simulate the enrichment analysis for the study 1
result1 <- data.frame(
  ID = paste0("geneset", 1:100),
  name = paste0("Pathway ", 1:100),
  description = paste0("Description ", 1:100),
  p.value = runif(100) / 10,
  pFDR = runif(100) / 5,
  size = runif(100, 100, 500),
  nDE = runif(100, 10, 100),
  score = runif(100, -2, 2),
  normalizedScore = runif(100)
)
rownames(result1) <- result1$ID

# Simulate the enrichment analysis for the study 2
result2 <- data.frame(
  ID = paste0("geneset", 1:100),
  name = paste0("Pathway ", 1:100),
  description = paste0("Description ", 1:100),
  p.value = runif(100) / 10,
  pFDR = runif(100) / 5,
  size = runif(100, 100, 500),
  nDE = runif(100, 10, 100),
  score = runif(100, -2, 2),
  normalizedScore = runif(100)
)
rownames(result2) <- result2$ID

# Store results from two studies in a list
resultsLst <- list(
  "study1" = result1,
  "study2" = result2
)

# Forest plot
plotForest(resultsLst, yAxis = "ID")
}
}
