% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-venn.R
\name{plotVennPathway}
\alias{plotVennPathway}
\title{Plot venndiagram from multiple pathway analysis results}
\usage{
plotVennPathway(PAResults, pThreshold = 0.05, useFDR = TRUE)
}
\arguments{
\item{PAResults}{A list of data frames with the results of pathway analysis.}

\item{pThreshold}{The p-value threshold to determine if a pathway is enriched.}

\item{useFDR}{Use the FDR adjusted p-value instead of the raw p-value.}
}
\description{
Plot a venndiagram from multiple pathway analysis results.
}
\examples{
#TODO add example
}
