% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volcano.R
\name{plotVolcanoDE}
\alias{plotVolcanoDE}
\title{Plot volcano plot from Pathway analysis results}
\usage{
plotVolcanoDE(
  DEResult,
  xAxis = "logFC",
  yAxis = c("-log10(pFDR)", "-log10(p.value)"),
  pThreshold = 0.05,
  statThreshold = 2
)
}
\arguments{
\item{xAxis}{The column to use for the x-axis.}

\item{yAxis}{The column to use for the y-axis.}

\item{pThreshold}{The p-value threshold to use for the horizontal line.}

\item{statThreshold}{The absolute value of the statistic threshold to use for the vertical line.}

\item{results}{A data frame with Pathway analysis results.
The columns are ID, name, description, p.value, pFDR, size, nDE, score and normalizedScore.}
}
\description{
Plot volcano plot from Pathway analysis results
}
