% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE-analysis.R
\name{runDEAnalysis}
\alias{runDEAnalysis}
\title{Differential expression analysis}
\usage{
runDEAnalysis(
  summarizedExperiment,
  method = c("limma", "DESeq2", "edgeR"),
  design,
  contrast,
  annotation = NULL
)
}
\arguments{
\item{summarizedExperiment}{SummarizedExperiment object}

\item{method}{Method to use for differential expression analysis. Can be "limma", "DESeq2" or "edgeR".}

\item{design}{A design model output by model.matrix.}

\item{contrast}{A contrast matrix. See limma::makeContrasts.}

\item{annotation}{A data frame mapping between probe IDs and entrez gene IDs.
If not provided, the function will try to get the mapping from the platform annotation in the SummarizedExperiment object.
If the annotation is not available, the function will return the probe IDs.
Regardless of the type of annotation, it must contains two columns: FROM and TO,
where FROM is the probe ID and TO is the entrez gene ID.}
}
\value{
A SummarizedExperiment object with DE analysis results appended to the rowData slot with the following columns:
\itemize{
\item{ID}{gene ID. If annotation is provided, this will be the entrez gene ID. Otherwise, it will be the probe ID.}
\item{logFC}{log2 fold change}
\item{p.value}{p-value from the DE analysis using the specified method}
\item{pFDR}{p-value adjusted for multiple testing using Benjamini-Hochberg method}
\item{statistic}{statistic from the DE analysis using the specified method.
For limma, this is the t-statistic.
For DESeq2, this is the Wald statistic.
For edgeR, this is the log fold change.}
\item{avgExpr}{
For limma, it is the average expression.
For DESeq2, it is the log base mean.
For edgeR, it is the log CPM.
}
\item{logFCSE}{standard error of the log fold change.}
\item{sampleSize}{sample size used for DE analysis.}
}
The assay slot will contain the input expression/count matrix,
and the rownames will be mapped to the gene IDs if annotation is found in the input SummarizedExperiment object
or in the annotation parameter.
Other slots will be the same as in the input SummarizedExperiment object.
}
\description{
This function performs differential expression analysis using either limma, DESeq2 or edgeR.
}
\examples{
\dontrun{
#Loading necessary libraries
library(hgu133plus2.db)
library(AnnotationDbi)
library(SummarizedExperiment)
library(limma)
library(RCPA)
# generate a random gene expression matrix
set.seed(123)
exprs <- round(matrix(2^abs(rnorm(1000, sd = 4)), nrow = 100, ncol = 10))
# Assign gene names
rownames(exprs) <- sample(keys(hgu133plus2.db, keytype = "PROBEID"), nrow(exprs), replace = FALSE)
# Assign sample names
colnames(exprs) <- paste0("sample", 1:10)
# Generate control and condition samples
controlSamples <- paste0("sample", 1:5)
conditionSamples <- paste0("sample", 6:10)
# Get colData
colData <- data.frame(
    row.names = colnames(exprs),
    group = factor(c(rep("control", 
    length(controlSamples)), 
    rep("condition", 
    length(conditionSamples)))),
    pair = factor(c(seq_along(controlSamples), seq_along(conditionSamples)))
)
# Construct summarizedExperiment object
summarizedExperiment <- SummarizedExperiment(
    assays = list(counts = exprs),
    colData = colData
)
# control vs condition
design <- model.matrix(~0 + group, data = colData)
contrast <- makeContrasts("groupcondition-groupcontrol", levels = design)
# Perform DE analysis
DERes <- runDEAnalysis(summarizedExperiment, 
method = "DESeq2", 
design, contrast, 
annotation = "GPL570")

# View DE analysis data frame
rowData(DERes)
}
}
