% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE-analysis.R
\name{runDEInternal}
\alias{runDEInternal}
\alias{.runLimma}
\alias{.runDESeq2}
\alias{.runEdgeR}
\title{Differential expression analysis}
\usage{
.runLimma(exprs, design, contrast)

.runDESeq2(exprs, design, contrast)

.runEdgeR(exprs, design, contrast)
}
\arguments{
\item{exprs}{Normalized expression matrix. Rows are genes and columns are samples.}

\item{design}{A design model output by model.matrix.}

\item{contrast}{A contrast matrix output by limma::makeContrasts.}
}
\value{
A data frame with DE analysis results. Must contain the following columns: PROBEID, p.value, statistic.
}
\description{
Functions to perform differential expression analysis
These functions are used internally by runDEAnalysis.
}
