% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-analysis.R
\name{runPathwayAnalysis}
\alias{runPathwayAnalysis}
\title{Pathway Enrichment Analysisbiological_process}
\usage{
runPathwayAnalysis(
  summarizedExprimentObj,
  genesets,
  method = "ORA",
  nperm = 1000
)
}
\arguments{
\item{summarizedExprimentObj}{The generated SummarizedExpriment object from DE analysis result.}

\item{genesets}{The genesets definition, ex. KEGG genesets from getGeneSets function.}

\item{method}{The pathway analsyis method, including ORA, fgsea, and GSA.}

\item{nperm}{The number of permutations to run pathway analysis.}
}
\value{
A dataframe of pathway analysis result
}
\description{
This function performs patwhay analysis using GSA method (GeneSet Analysis).
This function is used internally by runPathwayAnalysis.
}
\details{
This function is used internally by runPathwayAnalysis.
}
